+++ src/data/inputNumber.js
@@ -1,15 +1,12 @@
 import { MissionUtils } from "@woowacourse/mission-utils";
-import { TEXT } from "../common/text";
+import { TEXT } from "../common/index";
 
 const inputNumber = async () => {
   const number = await MissionUtils.Console.readLineAsync(
     `${TEXT.INPUT_NUMBER}`
   );
-  const numberIsValid = isValidnumber(number);
-  if (!numberIsValid) throw Error(TEXT.INPUT_ERROR);
+  if (isNaN(number)) throw Error(TEXT.INPUT_ERROR);
   return number;
 };
-
-const isValidnumber = (number) => !isNaN(number);
 
 export default inputNumber;
