+++ src/race/printResult.js
@@ -1,5 +1,5 @@
 import { MissionUtils } from "@woowacourse/mission-utils";
-import { TEXT } from "./common/text";
+import { TEXT } from "../common/index";
 
 const printResult = (winners) =>
   MissionUtils.Console.print(`${TEXT.WINNER_MESSAGE} : ${winners.join(", ")}`);
